software_standalone/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spotmicro-standalone"
version = "0.1.0"
description = "Standalone controller for the Spot Micro quadruped without ROS"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "SpotMicro Contributors"}]
dependencies = [
  "numpy>=1.24",
  "PyYAML>=6.0"
]

[project.optional-dependencies]
hardware = ["smbus2>=0.4"]
plot = ["matplotlib>=3.8"]
dev = ["pytest", "black", "ruff"]

[project.scripts]
spotmicro-teleop = "spotmicro_standalone.apps.teleop:main"

[tool.setuptools.packages.find]
where = ["spotmicro_standalone"]

